+++ rails/init.rb
@@ -1,2 +1,0 @@
-ActiveRecord::Base.send(:include, ActsAsTenant::ModelExtensions)
-ActionController::Base.extend ActsAsTenant::ControllerExtensions