+++ spec/acts_as_tenant/sidekiq_spec.rb
@@ -4,6 +4,7 @@ require "acts_as_tenant/sidekiq"
 
 describe ActsAsTenant::Sidekiq do
   after { ActsAsTenant.current_tenant = nil }
+
   let(:account) { Account.new(id: 1234) }
   let(:message) { {"acts_as_tenant" => {"class" => "Account", "id" => 1234}} }
 
@@ -51,13 +52,9 @@ describe ActsAsTenant::Sidekiq do
     end
   end
 
-  describe "Sidekiq configuration" do
-    describe "client configuration" do
-      it "includes ActsAsTenant client" do
-        expect(Sidekiq.client_middleware.exists?(ActsAsTenant::Sidekiq::Client)).to eq(true)
-      end
-    end
-
-    # unable to test server configuration
+  it "includes ActsAsTenant client middleware" do
+    expect(Sidekiq.client_middleware.exists?(ActsAsTenant::Sidekiq::Client)).to eq(true)
   end
+
+  # unable to test server configuration
 end