+++ spec/dummy/app/mailers/application_mailer.rb
@@ -1,0 +1,4 @@
+class ApplicationMailer < ActionMailer::Base
+  default from: "from@example.com", to: "to@example.com"
+  layout "mailer"
+end