+++ spec/dummy/app/models/account.rb
@@ -1,0 +1,4 @@
+class Account < ActiveRecord::Base
+  has_many :projects
+  has_many :global_projects
+end