+++ spec/dummy/app/models/article.rb
@@ -1,0 +1,3 @@
+class Article < ActiveRecord::Base
+  has_many :polymorphic_tenant_comments, as: :polymorphic_tenant_commentable
+end