+++ spec/dummy/app/models/custom_counter_cache_task.rb
@@ -1,0 +1,4 @@
+class CustomCounterCacheTask < ActiveRecord::Base
+  self.table_name = "projects"
+  acts_as_tenant(:account, counter_cache: "projects_count")
+end