+++ spec/dummy/app/models/custom_foreign_key_task.rb
@@ -1,0 +1,4 @@
+class CustomForeignKeyTask < ActiveRecord::Base
+  acts_as_tenant(:account, foreign_key: "accountID")
+  validates_uniqueness_to_tenant :name
+end