+++ spec/dummy/app/models/custom_primary_key_task.rb
@@ -1,0 +1,4 @@
+class CustomPrimaryKeyTask < ActiveRecord::Base
+  acts_as_tenant(:account, foreign_key: "name", primary_key: "name")
+  validates_presence_of :name
+end