+++ spec/dummy/app/models/manager.rb
@@ -1,0 +1,4 @@
+class Manager < ActiveRecord::Base
+  belongs_to :project
+  acts_as_tenant :account
+end