+++ spec/dummy/app/models/task.rb
@@ -1,0 +1,7 @@
+class Task < ActiveRecord::Base
+  belongs_to :project
+  default_scope -> { where(completed: nil).order("name") }
+
+  acts_as_tenant :account
+  validates_uniqueness_of :name
+end