+++ spec/dummy/app/models/unscoped_model.rb
@@ -1,0 +1,3 @@
+class UnscopedModel < ActiveRecord::Base
+  validates_uniqueness_of :name
+end